refs #5887 use a model so we do not have to expose the method in the API. I tried to write a test for this but was not able to since lots of IntegrationTest magic
